### 2023-05-22 #5483 @sbenthall
Hello @jstac ,

Before beginning to search for reviewers, I have a few points I wanted to bring up.

Can you please confirm that the scope of the submission is QuantEcon.py as developed here:
https://github.com/quantecon/QuantEcon.py

At the end of the review process, you will be asked to cut a release of the repository and publish it to an archive, such as Zenodo, so that it gets a DOI.

Most often, authors include their paper in the same repository as the software that is under review. (Perhaps in its own branch.) I wonder if you would consider consolidating your paper and the software code in some way.

The other thing I should ask is if you have any potential reviewers in mind for your submission. I draw your attention in particular to the [Conflict of Interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). Speaking truthfully, I understand the prominence of QuantEcon in your field. Your centrality may make finding unconflicted reviewers challenging. Guidance from you might ease the search process.
