### 2023-05-23 beec748
FIX: random.draw: Add `=None`
Fix #704
