### 2023-05-23 #5483 @oliviaguest
> ```
> Software report:
>
> github.com/AlDanial/cloc v 1.88  T=0.01 s (507.7 files/s, 112718.0 lines/s)
> -------------------------------------------------------------------------------
> Language                     files          blank        comment           code
> -------------------------------------------------------------------------------
> Markdown                         2            107              0            443
> TeX                              1             33              0            280
> YAML                             1              2              4             19
> -------------------------------------------------------------------------------
> SUM:                             4            142              4            742
> -------------------------------------------------------------------------------
>
>
> gitinspector failed to run statistical information for the repository
> ```

just in case it's not obvious, this failed because repo for code and paper must be the same
