### 2023-05-23 fc38e57
FIX: DOC: Remove `html4_writer = True`
