+++ docs/source/conf.py
@@ -161,8 +161,6 @@ else:  # Local build. Need to import rtd theme
   html_theme = "sphinx_rtd_theme"
   html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 
-html4_writer = True
-
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
