+++ joss_paper/paper.bib
@@ -1,0 +1,218 @@
+@book{miranda2004applied,
+  title={Applied computational economics and finance},
+  author={Miranda, Mario J and Fackler, Paul L},
+  year={2004},
+  publisher={MIT press}
+}
+
+@book{stewart2009probability,
+  title={Probability, Markov chains, queues, and simulation: the mathematical basis of performance modeling},
+  author={Stewart, William J},
+  year={2009},
+  publisher={Princeton university press}
+}
+
+@misc{nijenhuis1978combinatorial,
+  title={Combinatorial Algorithms, Ch. 5},
+  author={Nijenhuis, A and Wilf, HS},
+  year={1978},
+  publisher={Academic Press, New York}
+}
+
+@article{chiang2010structured,
+  title={STRUCTURED DOUBLING ALGORITHM FOR DISCRETE-TIME ALGEBRAIC RICCATI EQUATIONS WITH SINGULAR CONTROL WEIGHTING MATRICES},
+  author={Chiang, Chun-Yueh and Fan, Hung-Yuan and Lin, Wen-Wei},
+  journal={Taiwanese Journal of Mathematics},
+  volume={14},
+  number={3A},
+  pages={933--954},
+  year={2010},
+  publisher={Mathematical Society of the Republic of China},
+  doi={10.11650/twjm/1500405875}
+}
+
+@article{mclennan2005imitation,
+  title={From imitation games to Kakutani},
+  author={McLennan, Andrew and Tourky, Rabee},
+  journal={Manuscript, available at http://www. econ. umn. edu/mclennan/Papers/papers. html},
+  volume={37},
+  pages={41},
+  year={2005},
+  publisher={Citeseer}
+}
+
+@book{murty1988linear,
+  title={Linear complementarity, linear and nonlinear programming},
+  author={Murty, Katta G and Yu, Feng-Tien},
+  volume={3},
+  year={1988},
+  publisher={Citeseer}
+}
+
+@book{puterman2014markov,
+  title={Markov decision processes: discrete stochastic dynamic programming},
+  author={Puterman, Martin L},
+  year={2014},
+  publisher={John Wiley \& Sons}
+}
+
+@article{lagarias1998convergence,
+  title={Convergence properties of the Nelder--Mead simplex method in low dimensions},
+  author={Lagarias, Jeffrey C and Reeds, James A and Wright, Margaret H and Wright, Paul E},
+  journal={SIAM Journal on optimization},
+  volume={9},
+  number={1},
+  pages={112--147},
+  year={1998},
+  publisher={SIAM}
+}
+
+@article{singer2004efficient,
+  title={Efficient implementation of the Nelder--Mead search algorithm},
+  author={Singer, Sa{\v{s}}a and Singer, Sanja},
+  journal={Applied Numerical Analysis \& Computational Mathematics},
+  volume={1},
+  number={2},
+  pages={524--534},
+  year={2004},
+  publisher={Wiley Online Library},
+  doi={10.1002/anac.200410015}
+}
+
+@article{gao2012implementing,
+  title={Implementing the Nelder-Mead simplex algorithm with adaptive parameters},
+  author={Gao, Fuchang and Han, Lixing},
+  journal={Computational Optimization and Applications},
+  volume={51},
+  number={1},
+  pages={259--277},
+  year={2012},
+  publisher={Springer},
+  doi={10.1007/s10589-010-9329-3}
+}
+
+@article{border2004gauss,
+  title={The Gauss--Jordan and Simplex Algorithms},
+  author={Border, KC},
+  journal={California Institute of Technology, Division of the Humanities and Social Sciences},
+  year={2004}
+}
+
[…]
