+++ joss_paper/paper.md
@@ -1,0 +1,539 @@
+---
+title: 'QuantEcon.py: A community based Python library for quantitative economics'
+tags:
+  - Python
+  - Economics
+authors:
+  - name: Quentin Batista
+    affiliation: 5
+  - name: Chase Coleman
+    affiliation: 4
+  - name: Yuya Furusawa
+    affiliation: 6
+  - name: Shu Hu
+    affiliation: 1
+  - name: Smit Lunagariya
+    affiliation: 3
+  - name: Spencer Lyon
+    affiliation: 4
+  - name: Matthew McKay
+    affiliation: 1
+  - name: Daisuke Oyama
+    affiliation: 2
+  - name: Thomas J. Sargent
+    affiliation: 4
+  - name: Zejin Shi
+    affiliation: 7
+  - name: John Stachurski
+    affiliation: 1
+  - name: Pablo Winant
+    affiliation: 8
+  - name: Natasha Watkins
+    affiliation: 1
+  - name: Ziyue Yang
+    affiliation: 1
+  - name: Hengcheng Zhang
+    affiliation: 1
+affiliations:
+  - name: The Australian National University
+    index: 1
+  - name: University of Tokyo
+    index: 2
+  - name: Indian Institute of Technology (BHU), Varanasi
+    index: 3
+  - name: New York University
+    index: 4
+  - name: Massachusetts Institute of Technology
+    index: 5
+  - name: Crop.inc
+    index: 6
+  - name: The University of Arizona
+    index: 7
+  - name: ESCP Business School and Ecole Polytechnique
+    index: 8
+date: 13 December 2023
+bibliography: paper.bib
+---
+
+# Summary
+
+Economics traditionally relied on tractable mathematical models, diagrams,
+and simple regression methods to analyze and understand economic phenomena.
+However, in recent decades, economists have increasingly shifted towards more
+computationally challenging problems, involving large numbers of heterogeneous
+agents and complex nonlinear interactions.
+
+[QuantEcon.py](https://github.com/quantecon/QuantEcon.py) is an open-source
+Python library that helps to support this shift
+towards more computational intensive research in the field of economics.  First
+released in 2014, [QuantEcon.py](https://github.com/quantecon/QuantEcon.py)
+has been under continuous development for around
+9 years. The library includes a wide range of functions for economic analysis,
+including numerical methods, data visualization, estimation, and dynamic
+programming, implementing a number of fundamental algorithms used in
+high performance computational economics. In this article we review the key
+features of the the library.
+
+# Statement of Need
+
+Economists use a variety of economic, statistical and mathematical models as
+building blocks for constructing larger and more fully-featured models. Some of
+these are relatively unique to economics and finance. For example, many
+macroeconomic and financial models include a stochastic volatility component,
+since asset markets often exhibit bursts of volatility. Other building blocks
+involve optimization routines, such as firms that maximize present value given
+estimated time paths for profits and interest rates.  Firms modeled in this way
+are then plugged into larger models that contain households, banks and other
+economic agents.
+
+[QuantEcon.py](https://github.com/quantecon/QuantEcon.py) focuses on
+supplying building blocks for constructing economic
[…]
