### 2023-05-23 #5483 @Smit-create
Many thanks, @sbenthall and @oliviaguest.

> Can you please confirm that the scope of the submission is QuantEcon.py as developed here:
https://github.com/quantecon/QuantEcon.py

Yes, the submission is for `QuantEcon.py`.

> Most often, authors include their paper in the same repository as the software that is under review. (Perhaps in its own branch.) I wonder if you would consider consolidating your paper and the software code in some way.

I have created a branch in the `QuantEcon.py` repository: https://github.com/QuantEcon/QuantEcon.py/tree/joss_paper, and the paper can be found inside the folder `joss_paper`. I have also tested the paper build and that works fine.

Thanks again for the reviews.
