### 2023-05-25 1c0c214
Fix issue #691 (#698)
* remove broken links in python files

* remove broken links

* Updated broken links

* Update URL

* Apply review changes

---------

Co-authored-by: HengCheng <79777246+2789372130@users.noreply.github.com>
