+++ quantecon/_kalman.py
@@ -4,7 +4,7 @@ Implements the Kalman filter for a linear Gaussian state space model.
 References
 ----------
 
-https://lectures.quantecon.org/py/kalman.html
+https://python.quantecon.org/kalman.html
 
 """
 from textwrap import dedent
@@ -53,11 +53,10 @@ class Kalman:
     K_infinity : array_like or scalar(float)
         The stationary Kalman gain.
 
-
     References
     ----------
 
-    https://lectures.quantecon.org/py/kalman.html
+    https://python.quantecon.org/kalman.html
 
     """
 
@@ -154,6 +153,7 @@ class Kalman:
         whitened_lss : LinearStateSpace
             This is the linear state space system that represents
             the whitened system
+
         """
         K = self.K_infinity
 
@@ -289,6 +289,7 @@ class Kalman:
         coeff_type : string, either 'ma' or 'var' (default='ma')
             The type of coefficent sequence to compute.  Either 'ma' for
             moving average or 'var' for VAR.
+            
         """
         # == simplify notation == #
         A, G = self.ss.A, self.ss.G
