### 2023-05-25 e359066
DOC: Pin sphinx<=6.2.1
