+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This is a minor release with bug fixes and improvements to documentation.
 
 1. [Fix for #704](https://github.com/QuantEcon/QuantEcon.py/commit/beec748020bace3834bff9afa83067f668357a24)
 
+Thank you to [oyamad](https://github.com/oyamad), [HengchengZhang](https://github.com/HengchengZhang) for your
+contributions
+
 ## Ver 0.7.0 (24-April-2023)
 
 ### New
