+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
-## Ver 0.7.0 (24-February-2023)
+## Ver 0.7.1 (29-May-2023)
+
+This is a minor release with bug fixes and improvements to documentation. 
+
+### Fix
+
+1. [Fix for #704](https://github.com/QuantEcon/QuantEcon.py/commit/beec748020bace3834bff9afa83067f668357a24)
+
+Thank you to [oyamad](https://github.com/oyamad), [HengchengZhang](https://github.com/HengchengZhang) for your
+contributions
+
+## Ver 0.7.0 (24-April-2023)
 
 ### New
 
