### 2023-05-31 #5483 @jstac
@sbenthall Many thanks for your comments and my apologies for the delay in replying!

I confirm that the scope of the submission is QuantEcon.py as developed here:
https://github.com/quantecon/QuantEcon.py and the repository has now been changed to reflect this.

Regarding potential referees, here is a very random selection of economists I don't know.  I've never met or conversed with any of these people and I'm not aware of connections between them and QuantEcon.  But they are all involved in quantitative economics and have GH accounts.

* Mridul Seth (https://github.com/MridulS)
* Alan Lujan (https://quantmacro.org/)
* Regi Kusumaatmadja (https://github.com/abdurrahmanregi)
* Gerome Wolf (https://github.com/gerwolf)
* Rafael Araujo (https://github.com/RafaelCRAraujo)

Someone in the open source econ community I have talked to but don't have strong connections with is Janoś Gabler (https://github.com/janosg).
