### 2023-06-22 #5585 @editorialbot
[REVIEW]: QuantEcon.py: A community based Python library for quantitative economics
**Submitting author:** @jstac (<a href="http://orcid.org/0000-0001-6716-0111">John Stachurski</a>)
**Repository:** https://github.com/QuantEcon/QuantEcon.py
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 0.7.1
