### 2023-07-05 QuantEcon/QuantEcon.py#711 @janosg
License mismatch between repo and JOSS paper
In your [joss draft](https://github.com/openjournals/joss-papers/blob/joss.05585/joss.05585/10.21105.joss.05585.pdf) (line 46) you say that QuantEcon.py is released under a MIT license. However, the [LICENSE file](https://github.com/QuantEcon/QuantEcon.py/blob/main/LICENSE)  is BSD-3.

openjournals/joss-reviews#5585
