### 2023-07-06 QuantEcon/QuantEcon.py#711 @mmcky
Thanks @janosg

The following PR https://github.com/QuantEcon/QuantEcon.py/pull/712 will address this, just seeking approval from all lead developers and then it will be merged.
