### 2023-07-12 QuantEcon/QuantEcon.py#711 @mmcky
fixed by #712
