### 2023-07-12 QuantEcon/QuantEcon.py#714 @mmcky
ENH: Example Gallery
The [old example gallery](https://github.com/quantecon/QuantEcon.notebooks) is outdated and should be archived.

The [QuantEcon Lectures](https://quantecon.org/projects/#filter=lecture) provides lots of examples, and we do have examples in the documentation (i.e. docstrings), perhaps we can construct a demo gallery using those examples to be more directly accessible?
