### 2023-07-12 #5585 @mmcky
@janosg thank you for the issues and review. Greatly appreciated. I think I have now addressed both comments linked above. I have also opened [this discussion](https://github.com/QuantEcon/QuantEcon.py/issues/714) to see if we can make improvements through a more direct example gallery. Currently examples are provided across our lecture series and in the QuantEcon.py documentation docstrings.
