### 2023-07-12 QuantEcon/QuantEcon.py#714 @janosg
If you are thinking about a general re-organization of the documentation, I would like to suggest [this system](https://documentation.divio.com/) which I found very helpful in my own projects. It groups documentation in four categories:

- Reference: This is helpful for experienced users who want to look up a specific syntax. The current documentation is almost exclusively written in reference style (auto-generated from docstrings)
- Explanation: This is helpful for people who want to learn the economics/numerics and syntax at the same time. This is the style of the quantecon lectures.
- Tutorials: Tutorials help newbies to discover the functionality of a package. In my experience, tutorials can increase the adoption of a package. The submitted paper is written in tutorial style, Having a notebook with all examples from the paper would be a great getting started tutorial.
- How-To Guides: How-To Guides show how to achieve a very specific task without explaining any background that is not immediately relevant. I think this is not covered in the current documentation.

A great example of documentation that uses this system is [pytask](https://pytask-dev.readthedocs.io/en/stable/#)
