### 2023-07-12 #5585 @janosg
Thanks @mmcky. I added a suggestion to your discussion but that is not a prerequisite for acceptance.

quantecon.py is a great package and it was a pleasure to review it!

@sbenthall my review is complete.
