### 2023-07-13 QuantEcon/QuantEcon.py#714 @mmcky
thanks @janosg -- nice suggestion.
