### 2023-10-30 #5585 @mnwhite
It was probably long ago. I am the holdup. I began working on my review,
then got distracted, and now I have a newborn. I will try to get to this
ASAP, I'm really sorry, John.

On Mon, Oct 30, 2023, 12:08 AM John Stachurski ***@***.***>
wrote:

> @arfon <https://github.com/arfon> @sbenthall
> <https://github.com/sbenthall> Is there a target date for the first round
> referee process? Please let us know if we can help expedite by providing
> more information.
>
