### 2023-10-30 34560b8
Removing randomness from lqnash
