### 2023-11-21 #5585 @mnwhite
"You're hurting young economists" is my trigger. I will get this done
tomorrow.

On Mon, Nov 20, 2023, 11:43 PM John Stachurski ***@***.***>
wrote:

> @mnwhite <https://github.com/mnwhite> Sorry to keep pinging you but it
> would help some of the young people listed as coauthors. Hopefully it won't
> take too long.
>
