### 2023-11-25 #5585 @mnwhite
Too late on the cookies / you can't stop me! It's too much for one person
to eat anyway, so give to hungry students.

On Sat, Nov 25, 2023, 3:46 PM John Stachurski ***@***.***>
wrote:

> Thanks for your time @mnwhite <https://github.com/mnwhite> No need for
> cookies --- I remember what it was like to have a new baby in the house.
> And lots of people are waiting on me for various tasks, so I would feel bad
> if everyone else was on time :)
>
