### 2023-11-27 #5585 @mnwhite
Awesome, it was the least I could do after the long delay. Good luck to any
of your students on the market!

On Sun, Nov 26, 2023, 7:54 PM John Stachurski ***@***.***>
wrote:

> Sweets and cookies have landed, and look amazing, thanks @mnwhite
> <https://github.com/mnwhite> :-) You have made the hungry students (and
> admin staff) very happy.
>
