### 2023-11-27 #5585 @sbenthall
@jstac Could you please title-case your references by, for example, writing them out like so in the bibtex file:

`An {E}xperimental {A}nalysis of {L}emke-{H}owson {A}lgorithm`

In general, this is getting to be the last chance to make edits.
I'll be doing a proof-reading pass this week.
If you could double-check the items on the checklist, that would be great.

When this is finalized (should be by the end of the week!), the last steps are to tag a release, publish an archive online to get a DOI, and then we wrap up.
