### 2023-11-29 560d0ea
bibtex: title-case all titles
