+++ joss_paper/paper.bib
@@ -1,12 +1,12 @@
 @book{miranda2004applied,
-  title={Applied computational economics and finance},
+  title={Applied Computational Economics and Finance},
   author={Miranda, Mario J and Fackler, Paul L},
   year={2004},
   publisher={MIT press}
 }
 
 @book{stewart2009probability,
-  title={Probability, Markov chains, queues, and simulation: the mathematical basis of performance modeling},
+  title={Probability, Markov Chains, Queues, and Simulation: the Mathematical Basis of Performance Modeling},
   author={Stewart, William J},
   year={2009},
   publisher={Princeton university press}
@@ -20,7 +20,7 @@
 }
 
 @article{chiang2010structured,
-  title={STRUCTURED DOUBLING ALGORITHM FOR DISCRETE-TIME ALGEBRAIC RICCATI EQUATIONS WITH SINGULAR CONTROL WEIGHTING MATRICES},
+  title={Structured Doubling Algorithm for Discrete-Time Algebraic Riccati Equations with Singular Control Weighting Matrices},
   author={Chiang, Chun-Yueh and Fan, Hung-Yuan and Lin, Wen-Wei},
   journal={Taiwanese Journal of Mathematics},
   volume={14},
@@ -32,7 +32,7 @@
 }
 
 @article{mclennan2005imitation,
-  title={From imitation games to Kakutani},
+  title={From Imitation Games to Kakutani},
   author={McLennan, Andrew and Tourky, Rabee},
   journal={Manuscript, available at http://www. econ. umn. edu/mclennan/Papers/papers. html},
   volume={37},
@@ -42,7 +42,7 @@
 }
 
 @book{murty1988linear,
-  title={Linear complementarity, linear and nonlinear programming},
+  title={Linear Complementarity, Linear and Nonlinear Programming},
   author={Murty, Katta G and Yu, Feng-Tien},
   volume={3},
   year={1988},
@@ -50,14 +50,14 @@
 }
 
 @book{puterman2014markov,
-  title={Markov decision processes: discrete stochastic dynamic programming},
+  title={Markov Decision Processes: Discrete Stochastic Dynamic Programming},
   author={Puterman, Martin L},
   year={2014},
   publisher={John Wiley \& Sons}
 }
 
 @article{lagarias1998convergence,
-  title={Convergence properties of the Nelder--Mead simplex method in low dimensions},
+  title={Convergence Properties of the Nelder--Mead Simplex Method in Low Dimensions},
   author={Lagarias, Jeffrey C and Reeds, James A and Wright, Margaret H and Wright, Paul E},
   journal={SIAM Journal on optimization},
   volume={9},
@@ -68,7 +68,7 @@
 }
 
 @article{singer2004efficient,
-  title={Efficient implementation of the Nelder--Mead search algorithm},
+  title={Efficient Implementation of the Nelder--Mead Search Algorithm},
   author={Singer, Sa{\v{s}}a and Singer, Sanja},
   journal={Applied Numerical Analysis \& Computational Mathematics},
   volume={1},
@@ -80,7 +80,7 @@
 }
 
 @article{gao2012implementing,
-  title={Implementing the Nelder-Mead simplex algorithm with adaptive parameters},
+  title={Implementing the Nelder-Mead Simplex Algorithm with Adaptive Parameters},
   author={Gao, Fuchang and Han, Lixing},
   journal={Computational Optimization and Applications},
   volume={51},
@@ -99,14 +99,14 @@
 }
 
 @article{codenotti2008,
-  title={An experimental analysis of lemke-howson algorithm},
[…]
