### 2023-11-29 #5585 @mmcky
> Could you please title-case your references

Thanks @sbenthall I have just pushed updates to the bibtex file to add title-case.

- [ ] I will also do a final read through and run through the checklist mentioned above tomorrow.
