### 2023-11-30 #5585 @mmcky
thanks @sbenthall I have made some minor typo adjustments to the manuscript and edited the `bibles` file.

re: `bibtex` file I changed the titles to `title-case` but did you want the `brackets` as shown in the code-block?

```
An {E}xperimental {A}nalysis of {L}emke-{H}owson {A}lgorithm
```

I interpret those as the letters that needed to change in the bibles file (rather than the syntax you'd like used -- but happy to update tomorrow if needed).
