### 2023-11-30 #5585 @sbenthall
@mmcky The curly braces in BibTex are a signal to the renderer to use literal capitalization. Otherwise, it reverts those letters to lower case. In the rendered PDF, those letters are still lower case.

![image](https://github.com/openjournals/joss-reviews/assets/68752/6af6fc00-46fc-446b-9f79-2f4a58315c54)
