+++ joss_paper/paper.bib
@@ -1,26 +1,26 @@
 @book{miranda2004applied,
-  title={Applied Computational Economics and Finance},
+  title={{A}pplied {C}omputational {E}conomics and {F}inance},
   author={Miranda, Mario J and Fackler, Paul L},
   year={2004},
   publisher={MIT press}
 }
 
 @book{stewart2009probability,
-  title={Probability, Markov Chains, Queues, and Simulation: the Mathematical Basis of Performance Modeling},
+  title={{P}robability, {M}arkov {C}hains, {Q}ueues, and {S}imulation: the {M}athematical {B}asis of {P}erformance {M}odeling},
   author={Stewart, William J},
   year={2009},
   publisher={Princeton university press}
 }
 
 @misc{nijenhuis1978combinatorial,
-  title={Combinatorial Algorithms, Ch. 5},
+  title={{C}ombinatorial {A}lgorithms, Ch. 5},
   author={Nijenhuis, A and Wilf, HS},
   year={1978},
   publisher={Academic Press, New York}
 }
 
 @article{chiang2010structured,
-  title={Structured Doubling Algorithm for Discrete-Time Algebraic Riccati Equations with Singular Control Weighting Matrices},
+  title={{S}tructured {D}oubling {A}lgorithm for {D}iscrete-{T}ime {A}lgebraic {R}iccati {E}quations with {S}ingular {C}ontrol {W}eighting {M}atrices},
   author={Chiang, Chun-Yueh and Fan, Hung-Yuan and Lin, Wen-Wei},
   journal={Taiwanese Journal of Mathematics},
   volume={14},
@@ -32,9 +32,9 @@
 }
 
 @article{mclennan2005imitation,
-  title={From Imitation Games to Kakutani},
+  title={{F}rom {I}mitation {G}ames to {K}akutani},
   author={McLennan, Andrew and Tourky, Rabee},
-  journal={Manuscript, available at http://www. econ. umn. edu/mclennan/Papers/papers. html},
+  journal={Manuscript, available at http://www.econ.umn.edu/mclennan/Papers/papers.html},
   volume={37},
   pages={41},
   year={2005},
@@ -42,7 +42,7 @@
 }
 
 @book{murty1988linear,
-  title={Linear Complementarity, Linear and Nonlinear Programming},
+  title={{L}inear {C}omplementarity, {L}inear and {N}onlinear {P}rogramming},
   author={Murty, Katta G and Yu, Feng-Tien},
   volume={3},
   year={1988},
@@ -50,14 +50,14 @@
 }
 
 @book{puterman2014markov,
-  title={Markov Decision Processes: Discrete Stochastic Dynamic Programming},
+  title={{M}arkov {D}ecision {P}rocesses: {D}iscrete {S}tochastic {D}ynamic {P}rogramming},
   author={Puterman, Martin L},
   year={2014},
   publisher={John Wiley \& Sons}
 }
 
 @article{lagarias1998convergence,
-  title={Convergence Properties of the Nelder--Mead Simplex Method in Low Dimensions},
+  title={{C}onvergence {P}roperties of the {N}elder--{M}ead {S}implex {M}ethod in {L}ow {D}imensions},
   author={Lagarias, Jeffrey C and Reeds, James A and Wright, Margaret H and Wright, Paul E},
   journal={SIAM Journal on optimization},
   volume={9},
@@ -68,7 +68,7 @@
 }
 
 @article{singer2004efficient,
-  title={Efficient Implementation of the Nelder--Mead Search Algorithm},
+  title={{E}fficient {I}mplementation of the {N}elder--{M}ead {S}earch {A}lgorithm},
[…]
