### 2023-12-02 #5585 @mmcky
> It looks like you have listed a CC-BY license on the paper, but the repository has an MIT license.

Roger that.

I think the confusion was that the paper itself was to be `CC-BY` -- I will look into this as the `paper.md` file has

```
[QuantEcon.py](https://github.com/quantecon/QuantEcon.py) is released under the open-source MIT License
```

in the actual text.
