### 2023-12-02 #5585 @Smit-create
I am not exactly sure if we can configure the License text generated in the document (I still need to check more details to confirm). I think all the papers have the same License text: https://github.com/search?q=repo%3Aopenjournals%2Fjoss-papers%20%3Ccopyright-statement%3E&type=code ? For example, #5996 also has a MIT license but the paper doesn't https://joss.theoj.org/papers/10.21105/joss.05996.
