### 2023-12-03 #5585 @Smit-create
The License in the paper is coming from JOSS side which requires a CC-BY. See: https://github.com/openjournals/joss/issues/1215#issuecomment-1513548324. So, MIT license is of QuantEcon library but CC-BY is the license of the paper. Please feel free to correct me.
