### 2023-12-04 c939286
update article to techreport for better referencing
