+++ joss_paper/paper.bib
@@ -31,14 +31,11 @@
   doi={10.11650/twjm/1500405875}
 }
 
-@article{mclennan2005imitation,
+@techreport{mclennan2005imitation,
   title={{F}rom {I}mitation {G}ames to {K}akutani},
   author={McLennan, Andrew and Tourky, Rabee},
-  journal={Manuscript, available at http://www.econ.umn.edu/mclennan/Papers/papers.html},
-  volume={37},
-  pages={41},
+  institution={University of Pennsylvania},
   year={2005},
-  publisher={Citeseer}
 }
 
 @book{murty1988linear,
