+++ joss_paper/paper.bib
@@ -131,16 +131,6 @@
   year={2005}
 }
 
-@article{mclennan2005imitation,
-  title={{F}rom {I}mitation {G}ames to {K}akutani},
-  author={McLennan, Andrew and Tourky, Rabee},
-  journal={Manuscript, available at http://www. econ. umn. edu/mclennan/Papers/papers. html},
-  volume={37},
-  pages={41},
-  year={2005},
-  publisher={Citeseer}
-}
-
 @inproceedings{lam2015numba,
   title={{N}umba: {A} {LLVM}-based {P}ython {JIT} {C}ompiler},
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
