### 2023-12-04 fbda254
bibtex: add note for manuscript url
