+++ joss_paper/paper.bib
@@ -36,6 +36,7 @@
   author={McLennan, Andrew and Tourky, Rabee},
   institution={University of Pennsylvania},
   year={2005},
+  note = "Also available as \url{https://citeseerx.ist.psu.edu/document?repid=rep1&type=pdf&doi=77efd03ba49d1d035a79a9771792434882e3284d}"
 }
 
 @book{murty1988linear,
