### 2023-12-06 #5585 @mmcky
@sbenthall thanks for all the review and guidance.

I had another question with regards to the Zenodo upload.

> Archive the release on Zenodo/figshare/etc and post the DOI here.

Should we:
1. upload the paper only
2. upload the paper + copy of the `quantecon` package

I am currently assuming `#1` as the `quantecon==0.7.1` documents the software version.
