### 2023-12-06 #5585 @sbenthall
You know, I may be in error about the licensing issue. I apologize for the confusion; it's actually the first time I've noticed this requirement. (In the past, I haven't used the editoralbot checklist feature.)

The requirement, I believe, is that the license on the Zenodo archive is to be the same as that on the software. The paper remains CC; don't bother changing that.

The answer to your question @mmcky is actually #2. The issue is that GitHub and PyPI are not academic archives. Part of the point of JOSS is creative academic archives of software source code. So a package import is not sufficient.

I defer to @arfon or @oliviaguest on these interpretations.
