### 2023-12-06 #5585 @danielskatz
A couple of quick comments from me as another track editor...

What we need is the software to be uploaded to the repository, so that it can be archived permanently, which neither GitHub not PyPI do - the issue is the permanence of the archive. We don't need the paper source to be in the archive, though it can, as the paper will be archived through the journal.

The license for the software can be any OSI-approved license. The license for the paper is CC-BY.  The archive license should match the software license, since that's what the archive will mostly contain.
