### 2023-12-07 #5585 @mmcky
thanks @sbenthall and @danielskatz. I appreciate the clarifications and guidance.
I will get the final items on the checklist completed and then notify via this issue.
