### 2023-12-11 #5585 @mmcky
thanks again @danielskatz and @sbenthall.

I have now setup a [Zenodo record](https://zenodo.org/records/10345102) of the `quantecon==0.7.1` release.

![](https://zenodo.org/badge/DOI/10.5281/zenodo.10345102.svg)
