+++ .github/workflows/ci.yml
@@ -20,7 +20,7 @@ jobs:
         python-version: [3.9, '3.10', 3.11]
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@latest
       with:
         fetch-depth: 0
 
@@ -79,7 +79,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout source
-        uses: actions/checkout@v3
+        uses: actions/checkout@latest
       - name: Set up Python
         uses: actions/setup-python@v3
         with:
