### 2024-01-06 #5585 @oliviaguest
> Hi @sbenthall , happy new year and thanks again for your help in shepharding this paper through!
>
> Just quickly: one of the junior coauthors is applying for a scholarship and it would be very helpful if this was published on the main site. Is there a timeline? (I'm writing a recommendation letter for him so if you can let me know I'll adjust it accordingly.)
>
> Thanks in advance.

You can definitely use the DOI to refer to this now, if needed, see above. Apologies for slow reply; Xmas, new year, etc. Good luck with the application.
