### 2024-01-07 6df15c5
add bibtex for citation
