### 2024-01-15 7fffa18
Add citation (#719)
* add citation file

* update license year

* add bibtex for citation

* Update README.md

Co-authored-by: Daisuke Oyama <Oyama.Daisuke@gmail.com>

---------

Co-authored-by: Daisuke Oyama <Oyama.Daisuke@gmail.com>
