### 2024-01-15 d7c2f52
Update ci.yml to latest version (#718)
* Update ci.yml to latest version

* use v4
