### 2024-01-30 #5585 @oliviaguest
Huge thanks to the reviewers @janosg, @mnwhite and editor @sbenthall! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to the authors! 🥳 🍾
